requirements.txt
streamlit==1.42.0
openai==1.65.0
tavily-python==0.5.1
python-dotenv==1.0.1
